Introduce `PartialCommand` for better rendering/error information.
Now that the designator is dependant upon the table that the command
is being used from, rather than the command description, that
information needs to be included somewhere, so we introduced the
idea of a partial command that just includes the description and
the designators that were used to get that description.

That's all.