Introduce PresentationSchema & TopPresentationSchema.
Previously, our parameter descriptions used a `validator` method on
the presentation type directly in order to determine if an argument
was a valid match for the parameter. This worked, but presented
problems.

The first problem was that the validator operated on the object type
of a given presentation, and not the presentation itself.

This means that if you had say, a presentation type for a string
for a person's name. And you had a different presentation type
for a string for a place name, then the validator would only
be able to check against the JS type string and not the presentation
type.

The second problem was that there was no way to create a validator
for every single known presentation type.

So we fixed these problems by introducing a presentation schema to
describe what presentation types are accepted by a parameter.