Ok I've got it.
So two things were problematic.

The first is that I didn't understand that there's no partial
inference for type parameters. You either provide no type arguments
and get inference, or you provide them all, or you provide
some and there is no inference and the rest get defaulted if there
are any.

The next thing seems to be that whatever we were doing with `CommandMeta`
on `describeCommand` doesn't work. All the type parameters need to be
specified for them to be inferred. I don't really understand why yet,
but it's probably ok.