Cleanup `KeywordPropertyDescription`.
We now take the name from the associated property and close up weird
semantics with presentation types and flags by making the keyword
acceptor the top presentation schema.

We assume that the parser doesn't actually accept any associated
argument when a flag is used. But to be able to test that we need
to be able to run the commands in the unit tests.

This will require faking the command dispatcher or interface adaptor
and then as a pre-requisite probably putting all the dependencies
in some eays to use object that the client (user of the library)
will implement/extend some base thing that logs to console.