Update prefix extractor to allow transforming prefix.
In support of
https://github.com/the-draupnir-project/Draupnir/issues/678.

Introduce CommandNormaliser instead of PrefixExtractor.

The PrefixExtractor couldn't handle complicated behaviour,
such as using `!` as a prefix along side a command desginator.

Such as `!ban`.

It also was unable to handle display name mentions, mention pills. So
we have replaced it with a transformer, the command normaliser.